[refactor] tests
